Drop molecule code
The container-based setups can produce wildly different outcomes
to the full-fledged VM. As a result, most of the time is spent
debugging container-specific workarounds instead of the configuration
itself.

To that end, molecule test will not be provided for most of the roles.
Basic testing could be achieved via idempotence.
Other testing approaches may be considered in the future.